pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "miblib"
authors = [
    { name = "Michael Bungenstock", email = "sommerlich.ausrichtet.0l@icloud.com" },
]
description = "miblib is a collection of usefull Python functions"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["python", "gcp"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.26.0",
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = { attr = "mib.__version__" }

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
extend-select = ["N", "B", "E", "Q", "UP", "RUF", "SIM", "C90"] # D
